small merges from SAMBA_3_0; mostly typos, renames, etc...

+++ source/nsswitch/winbindd_pam.c
@@ -185,13 +185,13 @@ enum winbindd_result winbindd_pam_auth_crap(struct winbindd_cli_state *state)
 		goto done;
 	}
 
-        if (pull_utf8_talloc(mem_ctx, &user, state->request.data.auth_crap.user) == -1) {
+        if (pull_utf8_talloc(mem_ctx, &user, state->request.data.auth_crap.user) == (size_t)-1) {
 		DEBUG(0, ("winbindd_pam_auth_crap: pull_utf8_talloc failed!\n"));
 	}
 
 	if (*state->request.data.auth_crap.domain) {
 		char *dom = NULL;
-		if (pull_utf8_talloc(mem_ctx, &dom, state->request.data.auth_crap.domain) == -1) {
+		if (pull_utf8_talloc(mem_ctx, &dom, state->request.data.auth_crap.domain) == (size_t)-1) {
 			DEBUG(0, ("winbindd_pam_auth_crap: pull_utf8_talloc failed!\n"));
 		}
 		domain = dom;
@@ -215,7 +215,7 @@ enum winbindd_result winbindd_pam_auth_crap(struct winbindd_cli_state *state)
 
 	if (*state->request.data.auth_crap.workstation) {
 		char *wrk = NULL;
-		if (pull_utf8_talloc(mem_ctx, &wrk, state->request.data.auth_crap.workstation) == -1) {
+		if (pull_utf8_talloc(mem_ctx, &wrk, state->request.data.auth_crap.workstation) == (size_t)-1) {
 			DEBUG(0, ("winbindd_pam_auth_crap: pull_utf8_talloc failed!\n"));
 		}
 		workstation = wrk;
